feat: implement protocol fee manager
