+++ crates/precompiles/Cargo.toml
@@ -21,6 +21,7 @@ tempo-contracts = { workspace = true, default-features = false }
 tempo-precompiles = { workspace = true, default-features = false }
 tempo-precompiles-macros.workspace = true
 tempo-primitives = { workspace = true, default-features = false }
+tempo-revm = { workspace = true, default-features = false }
 
 # alloy
 alloy-consensus = { workspace = true, default-features = false }
